+++ inst/shiny/R/private.R
@@ -48,3 +48,12 @@ camelCaseToTitleCase <- function(string) {
   return(string)
 }
 
+addDays <- function(x, n) {
+  # Ensure that x is of class Date
+  xAsDate <- as.Date(x)
+  
+  # Add n days to xAsDate
+  newDate <- xAsDate + as.integer(n)
+  
+  return(newDate)
+}